Changes to be committed: 	modified:   src/cpp/chars.cpp 	modified:   src/cpp/cpp.cpp 	modified:   src/cpp/cut.cpp 	modified:   src/cpp/files.cpp 	modified:   src/cpp/global.cpp 	modified:   src/cpp/lines.cpp 	modified:   src/cpp/move.cpp 	modified:   src/cpp/nano.cpp 	modified:   src/cpp/prompt.cpp 	modified:   src/cpp/rcfile.cpp 	modified:   src/cpp/syntax.cpp 	modified:   src/cpp/text.cpp 	modified:   src/cpp/utils.cpp 	modified:   src/cpp/winio.cpp 	modified:   src/cpp/words.cpp 	modified:   src/include/definitions.h 	modified:   src/include/prototypes.h

+++ src/cpp/text.cpp
@@ -1588,12 +1588,12 @@ break_line(const char *textstart, long goal, bool snap_at_nl)
     /* If the whole line displays shorter than goal, we're done. */
     if ((long)column <= goal)
     {
-        return (pointer - textstart);
+        return (long)(pointer - textstart);
     }
     /* When wrapping a help text and no blank was found, force a line break. */
     if (snap_at_nl && lastblank == NULL)
     {
-        return step_left(textstart, pointer - textstart);
+        return (long)step_left(textstart, pointer - textstart);
     }
     /* If no blank was found within the goal width, seek one after it. */
     while (lastblank == NULL)
